Move HttpClientMock to an inner `mocks` package.
