+++ internal/anchorplatform/platform_api_service_test.go
@@ -10,10 +10,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stellar/stellar-disbursement-platform-backend/internal/serve/httpclient"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+
+	"github.com/stellar/stellar-disbursement-platform-backend/internal/serve/httpclient"
+	httpclientMocks "github.com/stellar/stellar-disbursement-platform-backend/internal/serve/httpclient/mocks"
 )
 
 func Test_NewAnchorPlatformAPIService(t *testing.T) {
@@ -73,7 +75,7 @@ func Test_NewAnchorPlatformAPIService(t *testing.T) {
 }
 
 func Test_updateAnchorTransactions(t *testing.T) {
-	httpClientMock := httpclient.HttpClientMock{}
+	httpClientMock := httpclientMocks.HttpClientMock{}
 	anchorPlatformAPIService, err := NewAnchorPlatformAPIService(&httpClientMock, "http://mock_anchor.com/", "jwt_secret_1234567890")
 	require.NoError(t, err)
 	ctx := context.Background()
@@ -159,7 +161,7 @@ func Test_updateAnchorTransactions(t *testing.T) {
 }
 
 func Test_PatchAnchorTransactionsPostRegistration(t *testing.T) {
-	httpClientMock := httpclient.HttpClientMock{}
+	httpClientMock := httpclientMocks.HttpClientMock{}
 	anchorPlatformAPIService, err := NewAnchorPlatformAPIService(&httpClientMock, "http://mock_anchor.com/", "jwt_secret_1234567890")
 	require.NoError(t, err)
 	ctx := context.Background()
@@ -227,7 +229,7 @@ func Test_PatchAnchorTransactionsPostRegistration(t *testing.T) {
 }
 
 func Test_PatchAnchorTransactionsPostSuccessCompletion(t *testing.T) {
-	httpClientMock := httpclient.HttpClientMock{}
+	httpClientMock := httpclientMocks.HttpClientMock{}
 	anchorPlatformAPIService, err := NewAnchorPlatformAPIService(&httpClientMock, "http://mock_anchor.com/", "jwt_secret_1234567890")
 	require.NoError(t, err)
 	ctx := context.Background()
@@ -332,7 +334,7 @@ func Test_PatchAnchorTransactionsPostSuccessCompletion(t *testing.T) {
 }
 
 func Test_PatchAnchorTransactionsPostErrorCompletion(t *testing.T) {
-	httpClientMock := httpclient.HttpClientMock{}
+	httpClientMock := httpclientMocks.HttpClientMock{}
 	anchorPlatformAPIService, err := NewAnchorPlatformAPIService(&httpClientMock, "http://mock_anchor.com/", "jwt_secret_1234567890")
 	require.NoError(t, err)
 	ctx := context.Background()
@@ -403,7 +405,7 @@ func Test_PatchAnchorTransactionsPostErrorCompletion(t *testing.T) {
 }
 
 func Test_getAnchorTransactions(t *testing.T) {
-	httpClientMock := httpclient.HttpClientMock{}
+	httpClientMock := httpclientMocks.HttpClientMock{}
 	apService, err := NewAnchorPlatformAPIService(&httpClientMock, "https://test.com/", "jwt_secret_1234567890")
 	require.NoError(t, err)
 	ctx := context.Background()
@@ -505,7 +507,7 @@ func Test_getAnchorTransactions(t *testing.T) {
 }
 
 func Test_IsAnchorProtectedByAuth(t *testing.T) {
-	httpClientMock := httpclient.HttpClientMock{}
+	httpClientMock := httpclientMocks.HttpClientMock{}
 	anchorPlatformAPIService, err := NewAnchorPlatformAPIService(&httpClientMock, "https://test.com/", "jwt_secret_1234567890")
 	require.NoError(t, err)
 	ctx := context.Background()
